refactor: move MyoClient to core, implement on_emg_data_aggregated()
